fixed language handling during import

+++ app/PersonioIntegration/Position.php
@@ -260,8 +260,8 @@ class Position {
 			// assign the position to its terms.
 			foreach ( Taxonomies::get_instance()->get_taxonomies() as $taxonomy_name => $taxonomy ) {
 				if ( ! empty( $taxonomy['attr']['rewrite']['slug'] ) ) {
-					// first remove all existing relations if list is appended.
-					if ( $taxonomy['append'] ) {
+					// first remove all existing relations if list is appended (but not for language).
+					if ( $taxonomy['append'] && $taxonomy_name !== WP_PERSONIO_INTEGRATION_TAXONOMY_LANGUAGES ) {
 						wp_delete_object_term_relationships( $this->get_id(), $taxonomy_name );
 					}
